refactor!: refactor the manager/ directory
This is the first step in the major refactor of the entire codebase.
This commit significantly simplifies the code of the event manager and
adds documentation to it.

BREAKING CHANGE: During the refactor, I removed some parts of the code
that were supposed to be workarounds for specific bugs, since I wasn't
able to reproduce them anymore. If the bugs arise again, I'll have to
add some of that code back.