fix: display windows correctly in overview
Previously, when opening the overview, the corner rounding was removed
from windows themselves and applied to window previews.

This prevented the windows from being slightly blurry, but after a
Mutter update, the code that was handling this broke, and window preview
sizes were calculated incorrectly.

This commit removes this behavior. It means that windows in overview
will be a bit pixelated, but it isn't very noticeable, and certainly
better than the windows being half transparent. Maybe the blurriness
issue will be easier to fix in the future, after I refactor the
codebase.

Fixes: #29