refactor!: refactor the utils directory
A third commit in a series of major refactors, this time it greatly
simplifies the utils directory, both in terms of code readability and in
terms of application logic.

Since the things defined inside of utils are used throughout the entire
codebase, this commit also introduces significant changes outside of the
utils directory.

BREAKING CHANGE: This commit changes the settings schema, which means
that a lot of the settings will have to be reset when the extension is
upgraded.

BREAKING CHANGE: The code that manually disabled shadows for X11 windows
has been removed, since it uses GNOME APIs which weren't supposed to be
public and which will are now removed[1]. I didn't notice any issues
after this change, but they may happen in the future.

[1]: https://gitlab.gnome.org/GNOME/mutter/-/commit/0104fbe577594bcc7fdc965be323be08e4bd35d0