refactor!: refactor the effect directory
Most of the actual code is left unchanged, since there wasn't much to
refactor. However, this commit adds a LOT of comments to the code of the
fragment shader, as well as some extra documentation in other places.

BREAKING CHANGE: GLSLEffect overrides the blending mode, which seemed to
cause issues in the past, so the mode had to be manually set to the
default value. However, removing this doesn't seem to cause any issues
anymore, so it's no longer done. Perhaps this breaks something in the
future though.