refactor: refactor extensions.ts
This is (should be) the last commit in the major project refactor. It
simplifies the code in the main `extnesion.ts` file, and moves the
monkey patch implementations to a separate directory.

Now that the refactor is complete, let me vent for a bit, for the 2
people in the entire world who read full commit messages for some
reason.

I started refactoring this project simply because I am a perfectionist
when it comes to code. I just can't stand having control over a repo
that doesn't have perfect, immediately readable code. And I don't really
regret doing this refactor. If I disappear for whatever reason, like
@yilozt did, this'll HOPEFULLY make it easier for someone else to take
over and create a new fork. Also, this refactor taught me a lot about
how GNOME shell and extensions work.

However, I also spent a lot of time on this. This project has been on my
GTD todo list, constantly reminding me that I have to work on this and
can't pick up anything new and exciting until I do, lest I abandon this.
I hate finishing project, and I loose interest really quickly, so this
may be way too much effort for something that probably won't pay off at
all. I kind of hate working on this now, which is the opposite of what
the refactor was supposed to achieve.

But also, this project has a lot of stars, and it makes me happy.

So yeah. I think it's fair to say that I have a love-hate relationship
with this extensions. But boy, I'm SO fucking happy that the refactor is
finally done.

Yeah, and if you encounter any bugs with this (that are not caused by
this commit), sorry; I plan to take a break from this pesky thing and
not touch it until the next GNOME release.